Remove write route display leakage
