+++ thingifier/src/main/java/uk/co/compendiumdev/thingifier/api/restapihandlers/RestApiDeleteHandler.java
@@ -38,6 +38,6 @@ public class RestApiDeleteHandler {
         }
 
         ThingCommandResult result = runtime.commandService(context).execute(mapping.getCommand());
-        return apiMapper.map(mapping.getCommand(), result);
+        return apiMapper.map(mapping, result);
     }
 }
