+++ thingifier/src/main/java/uk/co/compendiumdev/thingifier/api/restapihandlers/RestApiPutHandler.java
@@ -49,6 +49,6 @@ public class RestApiPutHandler {
         }
 
         ThingCommandResult result = runtime.commandService(context).execute(mapping.getCommand());
-        return apiMapper.map(mapping.getCommand(), result);
+        return apiMapper.map(mapping, result);
     }
 }
