+++ thingifier/src/main/java/uk/co/compendiumdev/thingifier/application/ApplicationExceptionMessages.java
@@ -1,0 +1,11 @@
+package uk.co.compendiumdev.thingifier.application;
+
+final class ApplicationExceptionMessages {
+
+    private ApplicationExceptionMessages() {}
+
+    static String messageFrom(final Exception exception) {
+        String message = exception.getMessage();
+        return message == null ? "" : message;
+    }
+}
