+++ thingifier/src/main/java/uk/co/compendiumdev/thingifier/application/command/DeleteThingCommand.java
@@ -4,13 +4,10 @@ public final class DeleteThingCommand implements ThingWriteCommand {
 
     private final String entityName;
     private final String identifier;
-    private final String routeDisplay;
 
-    public DeleteThingCommand(
-            final String entityName, final String identifier, final String routeDisplay) {
+    public DeleteThingCommand(final String entityName, final String identifier) {
         this.entityName = entityName;
         this.identifier = identifier;
-        this.routeDisplay = routeDisplay == null ? "" : routeDisplay;
     }
 
     public String getEntityName() {
@@ -19,9 +16,5 @@ public final class DeleteThingCommand implements ThingWriteCommand {
 
     public String getIdentifier() {
         return identifier;
-    }
-
-    public String getRouteDisplay() {
-        return routeDisplay;
     }
 }
